feat: add viewport audit Playwright spec and console demo seed
