+++ packages/db/src/index.ts
@@ -1,5 +1,6 @@
 export * from './blocks';
 export * from './client';
+export * from './console-demo';
 export * from './process';
 export * from './schema';
 export * from './skillsheet';
